Add structured logging as an alternative to stdout printing
This fixes #131 and adds a `oslog` handler for logging as an option.

Rather than change all logging to use this new endpoint by default I
left it as an option for the client to decide as perhaps people could be
relying on stdout printing.

The new variable is gated around the platforms which support this new
structured logging API (iOS 14 and similar).